job-orchestrator/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.1
aiosignal==1.3.1
asn1crypto==1.5.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloud-sql-python-connector==1.13.0
cryptography==43.0.3
Deprecated==1.2.15
Flask==3.1.0
frozenlist==1.5.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-pubsub==2.27.1
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
pg8000==1.31.2
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
scramp==1.4.5
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.16.0
yarl==1.17.1
zipp==3.21.0
